Use optimist for option parsing, add --stdout flag
Command-line options changed so that you can optionally specify a --renderer
or --parser instead of the --github-wiki/--json and --requirejs/--commonjs flags.

The --stdout flag causes each rendered documentation file to be printed to
standard out rather than written to a file. It is probably not useful except
with --renderer json.